+++ src/app/cart/cart.module.ts
@@ -4,7 +4,7 @@ import { CommonModule } from '@angular/common';
 import { CartListComponent } from './cart-list/cart-list.component';
 import { CartItemComponent } from './cart-item/cart-item.component';
 
-import { CartService } from './cart.service';
+import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
   declarations: [
@@ -12,13 +12,11 @@ import { CartService } from './cart.service';
     CartItemComponent
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    SharedModule
   ],
-  providers: [CartService],
   exports: [
-    CartListComponent,
-    CartItemComponent
+    CartListComponent
   ]
-
 })
 export class CartModule { }