+++ src/app/cart/cart.service.ts
@@ -1,10 +1,9 @@
 import { Injectable } from '@angular/core';
-
 import { Subject } from 'rxjs';
 
 import { Product } from '../product/product.model';
 
-let cart = [
+const cart = [
   new Product('cucumber', 'vegetable', true, 7, 1),
   new Product('dress', 'basic women item', true, 150, 1)
 ];
@@ -23,35 +22,32 @@ export class CartService {
   // Observable string streams
   public channel$ = this.channel.asObservable();
 
+  public totalQuantity = this.getQuantity();
+  public totalSum = this.getFullPrice();
 
-  getCartContent(): Array<Product> {
+
+  getCartContent(): Product[] {
     return cart;
   }
 
-  getQuantity() {
-    let quantity = 0;
-    cart.forEach((cartitem) => {
-      quantity += cartitem.quantity;
-    });
-
-    return quantity;
+  getQuantity(): number {
+    return cart.reduce((quantity, cartitem) => quantity += cartitem.quantity, 0);
   }
 
-  getFullPrice() {
-    let price = 0;
-    cart.forEach((cartitem) => {
-      price += cartitem.quantity * cartitem.price;
-    });
-
-    return price;
+  getFullPrice(): number {
+    return cart.reduce((price, cartitem) => price += cartitem.quantity * cartitem.price, 0);
   }
 
-  cleanCart() {
-    cart = [];
+  removeAllProducts(): void {
+    while (cart.length) {
+      cart.pop();
+    }
+
+    this.updateCartData();
     this.channel.next();
   }
 
-  addToCart(product) {
+  addProduct(product: Product): void {
     let existingProductIndex;
 
     cart.forEach((cartitem, i) => {
@@ -60,31 +56,32 @@ export class CartService {
       }
     });
 
-    existingProductIndex ? cart[existingProductIndex].quantity += 1 : cart.push(product);
+    existingProductIndex !== -1 ? cart[existingProductIndex].quantity += 1 : cart.push(product);
+
+    this.updateCartData();
     this.channel.next();
   }
 
-  removeProduct(product) {
-    cart.forEach((cartitem, i) => {
-      if (cartitem.name === product.name) {
-        cart.splice(i, 1);
-      }
-    });
+  removeProduct(product: Product): void {
+    const index = cart.findIndex((cartitem) => cartitem.name === product.name);
 
+    cart.splice(index, 1);
+    this.updateCartData();
     this.channel.next();
   }
 
-  changeProductQuantity(product, value) {
+  changeProductQuantity(product: Product, value: number): void {
     cart.forEach((cartitem, i) => {
       if (cartitem.name === product.name) {
-        if (value === 'sum') {
-          cartitem.quantity += 1;
-        } else if (value === 'sub') {
-          cartitem.quantity -= 1;
-        }
+        cartitem.quantity += value;
       }
     });
-
+    this.updateCartData();
     this.channel.next();
+  }
+
+  updateCartData(): void {
+    this.totalQuantity = this.getQuantity();
+    this.totalSum = this.getFullPrice();
   }
 }