+++ src/app/shared/config-options.service.ts
@@ -1,0 +1,23 @@
+import { Injectable } from '@angular/core';
+
+let config = {
+  id: '',
+  login: '',
+  email: ''
+};
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ConfigOptionsService {
+
+  constructor() { }
+
+  setConfig(object): void {
+    config = {...config, ...object};
+  }
+
+  getConfig(): {} {
+    return config;
+  }
+}