+++ src/app/shared/constant.service.ts
@@ -1,0 +1,6 @@
+const ConstantService = new Object({
+  App: 'TaskManager',
+  Ver: '1.0'
+});
+
+export default ConstantService;