+++ src/app/shared/generator-n.service.ts
@@ -1,0 +1,8 @@
+import { InjectionToken } from '@angular/core';
+import { GeneratorService } from './generator.service';
+
+export const NString = new InjectionToken<string>('NString');
+
+export function GeneratorFactory(n: number): Function {
+  return (generator: GeneratorService): string => generator.getString().substr(0, n);
+}