+++ src/app/shared/shared.module.ts
@@ -1,11 +1,25 @@
-import { NgModule } from '@angular/core';
+import { NgModule, InjectionToken } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { HoverElementDirective } from './hover-element.directive';
+import { FontClickDirective } from './font-click.directive';
+
+import { LocalStorageService } from './local-storage.service';
+import { ConfigOptionsService } from './config-options.service';
+import ConstantService from './constant.service';
+import { GeneratorService } from './generator.service';
+
+const ConstService = new InjectionToken<string>('ConstService');
 
 @NgModule({
-  declarations: [HoverElementDirective],
-  exports: [HoverElementDirective],
+  declarations: [HoverElementDirective, FontClickDirective],
+  exports: [HoverElementDirective, FontClickDirective],
+  providers: [
+    LocalStorageService,
+    ConfigOptionsService,
+    { provide: ConstService, useValue: ConstantService},
+    GeneratorService
+  ],
   imports: [
     CommonModule
   ]